+++ packages/studio-base/src/context/PanelCatalogContext.ts
@@ -2,11 +2,13 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
+import { isEmpty } from "lodash";
 import { ComponentType, createContext, useContext } from "react";
 
 import { PanelStatics } from "@foxglove/studio-base/components/Panel";
 import { ExtensionNamespace } from "@foxglove/studio-base/types/Extensions";
 import { PanelConfig } from "@foxglove/studio-base/types/panels";
+import { mightActuallyBePartial } from "@foxglove/studio-base/util/mightActuallyBePartial";
 
 export type PanelComponent = ComponentType<{ childId?: string; tabId?: string }> &
   PanelStatics<PanelConfig>;
@@ -51,6 +53,29 @@ export function usePanelCatalog(): PanelCatalog {
   }
 
   return panelCatalog;
+}
+
+// sanity checks to help panel authors debug issues
+export function verifyPanels(panels: readonly PanelInfo[]): void {
+  const panelTypes: Map<string, PanelInfo> = new Map();
+  for (const panel of panels) {
+    const { title, type, config } = mightActuallyBePartial(panel);
+    const dispName = title ?? type ?? "<unnamed>";
+    if (type == undefined || type.length === 0) {
+      throw new Error(`Panel component ${title} must declare a unique \`static panelType\``);
+    }
+    const existingPanel = mightActuallyBePartial(panelTypes.get(type));
+    if (existingPanel) {
+      const bothHaveEmptyConfigs = isEmpty(existingPanel.config) && isEmpty(config);
+      if (bothHaveEmptyConfigs) {
+        const otherDisplayName = existingPanel.title ?? existingPanel.type ?? "<unnamed>";
+        throw new Error(
+          `Two components have the same panelType ('${type}') and no preset configs: ${otherDisplayName} and ${dispName}`,
+        );
+      }
+    }
+    panelTypes.set(type, panel);
+  }
 }
 
 export default PanelCatalogContext;
